refactor: move auth into auth dir
